+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [IMPROVEMENT] Decorate network span kind as `client`. See [#1963][]
 - [FIX] Fix CPU spikes when Watchdog Terminations tracking is enabled. See [#1968][]
 - [FIX] Fix CPU spike when recording UITabBar using SessionReplay. See [#1967][]
+- [FEATURE] Add image privacy level configuration for Session Replay. See [#1959][]
 
 # 2.14.1 / 09-07-2024
 
@@ -743,6 +744,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1967]: https://github.com/DataDog/dd-sdk-ios/pull/1967
 [#1973]: https://github.com/DataDog/dd-sdk-ios/pull/1973
 [#1988]: https://github.com/DataDog/dd-sdk-ios/pull/1988
+[#1959]: https://github.com/DataDog/dd-sdk-ios/pull/1959
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu